refactor(analysis-types): split dependency DTOs
