+++ crates/tokmd-analysis-types/src/lib.rs
@@ -17,6 +17,7 @@
 mod api_surface;
 mod churn;
 mod corporate;
+mod dependencies;
 mod derived;
 mod duplication;
 mod effort;
@@ -34,6 +35,7 @@ use tokmd_types::{ScanStatus, ToolInfo};
 pub use api_surface::{ApiExportItem, ApiSurfaceReport, LangApiSurface, ModuleApiRow};
 pub use churn::{ChurnTrend, PredictiveChurnReport, TrendClass};
 pub use corporate::{CorporateFingerprint, DomainStat};
+pub use dependencies::{DependencyReport, LockfileReport};
 pub use derived::{
     BoilerplateReport, ContextWindowReport, DerivedReport, DerivedTotals, DistributionReport,
     FileStatRow, HistogramBucket, IntegrityReport, LangPurityReport, LangPurityRow, MaxFileReport,
@@ -57,7 +59,7 @@ pub use git::{
     ModuleIntentRow,
 };
 pub use license::{LicenseFinding, LicenseReport, LicenseSourceKind};
-pub use supply::{AssetCategoryRow, AssetFileRow, AssetReport, DependencyReport, LockfileReport};
+pub use supply::{AssetCategoryRow, AssetFileRow, AssetReport};
 pub use topics::{TopicClouds, TopicTerm};
 pub use util::{
     AnalysisLimits, empty_file_row, is_infra_lang, is_test_path, normalize_path, normalize_root,
