+++ crates/tokmd-analysis-types/src/supply.rs
@@ -1,7 +1,7 @@
-//! Asset and dependency receipt DTOs.
+//! Asset receipt DTOs.
 //!
-//! These supply-chain-adjacent contract types remain re-exported from the
-//! crate root to preserve existing `tokmd_analysis_types::...` names.
+//! These supply-chain-adjacent contract types remain re-exported from the crate
+//! root to preserve existing `tokmd_analysis_types::...` names.
 
 use serde::{Deserialize, Serialize};
 
@@ -27,17 +27,4 @@ pub struct AssetFileRow {
     pub bytes: u64,
     pub category: String,
     pub extension: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DependencyReport {
-    pub total: usize,
-    pub lockfiles: Vec<LockfileReport>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LockfileReport {
-    pub path: String,
-    pub kind: String,
-    pub dependencies: usize,
 }
